backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
Pillow==10.0.0
